--- a/code/frameworks/react-native-web-vite/src/typings.d.ts
+++ b/code/frameworks/react-native-web-vite/src/typings.d.ts
@@ -1,0 +1,1 @@
+declare module '@bunchtogether/vite-plugin-flow';