Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

Added support for flow which is used by some React Native libraries and React Native source code itself.

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-29756-sha-18442450`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-29756-sha-18442450 sandbox` or in an existing project with `npx storybook@0.0.0-pr-29756-sha-18442450 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-29756-sha-18442450`](https://npmjs.com/package/storybook/v/0.0.0-pr-29756-sha-18442450) |
| **Triggered by** | @shilman |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`dannyhw/fix-rnw-flow-support`](https://github.com/storybookjs/storybook/tree/dannyhw/fix-rnw-flow-support) |
| **Commit** | [`18442450`](https://github.com/storybookjs/storybook/commit/18442450017d77202dc9fd5aa4910af9c192fce4) |
| **Datetime** | Sat Nov 30 13:45:21 UTC 2024 (`1732974321`) |
| **Workflow run** | [12096885441](https://github.com/storybookjs/storybook/actions/runs/12096885441) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=29756`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  77.7 MB | 77.7 MB | 364 B | -0.37 | 0% |
| initSize |  130 MB | 130 MB | 11.4 kB | **-1.73** | 0% |
| diffSize |  52.4 MB | 52.4 MB | 11 kB | **-1.73** | 0% |
| buildSize |  6.83 MB | 6.83 MB | 0 B | **1.45** | 0% |
| buildSbAddonsSize |  1.51 MB | 1.51 MB | 0 B | 0.42 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.86 MB | 1.86 MB | 0 B | 1.21 | 0% |
| buildSbPreviewSize |  271 kB | 271 kB | 0 B | 0.35 | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.84 MB | 3.84 MB | 0 B | 1.01 | 0% |
| buildPreviewSize |  3 MB | 3 MB | 0 B | **1.73** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.6s | 23.4s | 16.7s | **1.27** | 🔺71.5% |
| generateTime |  19.8s | 18.2s | -1s -545ms | -0.85 | -8.4% |
| initTime |  13s | 11.8s | -1s -240ms | -0.88 | -10.5% |
| buildTime |  7.9s | 8.5s | 617ms | -0.47 | 7.2% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  4.2s | 4.9s | 725ms | -0.59 | 14.7% |
| devManagerResponsive |  3.2s | 3.5s | 379ms | -0.11 | 10.6% |
| devManagerHeaderVisible |  485ms | 596ms | 111ms | -0.04 | 18.6% |
| devManagerIndexVisible |  543ms | 624ms | 81ms | -0.19 | 13% |
| devStoryVisibleUncached |  1.6s | 1.7s | 184ms | 0.77 | 10.3% |
| devStoryVisible |  541ms | 622ms | 81ms | -0.15 | 13% |
| devAutodocsVisible |  419ms | 487ms | 68ms | -0.28 | 14% |
| devMDXVisible |  426ms | 559ms | 133ms | 0.06 | 23.8% |
| buildManagerHeaderVisible |  456ms | 587ms | 131ms | -0.03 | 22.3% |
| buildManagerIndexVisible |  469ms | 602ms | 133ms | -0.04 | 22.1% |
| buildStoryVisible |  456ms | 590ms | 134ms | 0 | 22.7% |
| buildAutodocsVisible |  384ms | 476ms | 92ms | -0.04 | 19.3% |
| buildMDXVisible |  368ms | 447ms | 79ms | -0.2 | 17.7% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Added Flow support to the React Native Web Vite framework in Storybook by integrating and configuring the vite-plugin-flow package for handling Flow types in React Native libraries.

- Added `@bunchtogether/vite-plugin-flow` dependency in `code/frameworks/react-native-web-vite/package.json`
- Configured Flow plugins in `code/frameworks/react-native-web-vite/src/preset.ts` with proper plugin ordering (Flow before React)
- Added TypeScript module declaration in `code/frameworks/react-native-web-vite/src/typings.d.ts`
- Note: Empty exclude array in Flow plugin configuration may need review for performance



<sub>💡 (1/5) You can manually trigger the bot by mentioning @greptileai in a comment!</sub>

<!-- /greptile_comment -->